Move all Iterator stuff into Collection; removes Iterator class
